% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ensbiascoR.example1}
\alias{ensbiascoR.example1}
\title{A test dataset to run example1 on http://ensbiascor.r-forge.r-project.org/ensbiascoR_example1.html}
\format{A list with two entries:
\describe{
  \item{obs.data}{Data frame with reanalysis data from ERA-Interim, aggregated over Central Europe (JJA), for the years 1979-2014.}
  \item{mod.data}{Data frame with HadRM3P simulated data, aggregated over Central Europe (JJA), for the years 1985-2010 and 800 ensemble members for each year.}
  ...
}}
\source{
Sippel, S., Otto, F. E. L., Forkel, M., Allen, M. R., Guillod, B. P., Heimann, M., Reichstein, M., Seneviratne, S. I., Thonicke, K. & Mahecha, M. D. (2016) A novel bias correction methodology for climate impact simulations. Earth System Dynamics, 7, 71-88. doi:10.5194/esd-7-71-2016.
}
\usage{
data(ensbiascoR.example1)
}
\description{
A test dataset to run example1 on http://ensbiascor.r-forge.r-project.org/ensbiascoR_example1.html
}
\examples{
data(ensbiascoR.example1)
}
\keyword{datasets}

