% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ensbiascoR.example2}
\alias{ensbiascoR.example2}
\title{A test dataset to run example2 on http://ensbiascor.r-forge.r-project.org/ensbiascoR_example2.html}
\format{A list with three locations: AT-Vienna (AT_WIE), DE-Jena (DE_JEN), and NL-De Bilt (NL_DEB). Each location contains a list with two entries described below:
\describe{
  \item{obs.data}{Data frame with station data (from the ECA&D dataset), processed to seasonal maxima of indiviual heat-health related metrics}
  \item{mod.data}{Data frame with HadRM3P simulated data, for the years 1985-2010, ~20.000 ensemble members in total, and processed to seasonal maxima of indiviual heat-health related metrics}
  ...
}}
\source{
Will follow soon.
}
\usage{
data(ensbiascoR.example2)
}
\description{
A test dataset to run example2 on http://ensbiascor.r-forge.r-project.org/ensbiascoR_example2.html
}
\examples{
data(ensbiascoR.example2)
}
\keyword{datasets}

