% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.percentile.transfer.function.R
\name{get.kernel.percentiles}
\alias{get.kernel.percentiles}
\title{A function to get percentiles of a kernel}
\usage{
get.kernel.percentiles(data, kernel.quantiles = seq(0, 1, 0.001))
}
\arguments{
\item{data}{A sample vector}

\item{kernel.quantiles}{The quantiles of the kernel to be computed}
}
\value{
A numeric vector that contains quantiles / percentiles of a fitted kernel
}
\description{
Get kernel quantiles for a given sample
}
\examples{
data(ensbiascor.example1)
obs.kernel = get.kernel.percentiles(data=ensbiascor.example1$obs.data$Tair[which(ensbiascor.example1$obs.data$Year \%in\% 1985:2010)])
mod.kernel = get.kernel.percentiles(data=ensbiascor.example1$mod.data$Tair)
plot(ecdf(ensbiascor.example1$obs.data$Tair[which(ensbiascor.example1$obs.data$Year \%in\% 1985:2010)]), bty='n', xlab = "Observational metric", ylab="Cumulative Density Function", xlim=c(14,25), main="", yaxt="n")
axis(side=2, las=1)
lines(x = obs.kernel, y=kernel.quantiles, col="blue", lwd=2)
lines(x = mod.kernel, y=kernel.quantiles, col="red", lwd = 2)
}
\author{
Sebastian Sippel
}
\references{
Sippel, S., Otto, F. E. L., Forkel, M., Allen, M. R., Guillod, B. P., Heimann, M., Reichstein, M., Seneviratne, S. I., Thonicke, K. & Mahecha, M. D. (2016) A novel bias correction methodology for climate impact simulations. Earth System Dynamics, 7, 71-88. doi:10.5194/esd-7-71-2016.
}

