% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.percentile.transfer.function.R
\name{get.percentile.transfer.function}
\alias{get.percentile.transfer.function}
\title{A function to derive a percentile transfer function}
\usage{
get.percentile.transfer.function(obs.grid.cell, mod.grid.cell, n.sample=10000)
}
\arguments{
\item{obs.grid.cell}{A numeric vector of observed values}

\item{mod.grid.cell}{A numeric vector of simulated values}

\item{n.sample}{Sample size}
}
\value{
XYZ
}
\description{
Calculates a transfer function between percentile of observed / simulated distributions
}
\details{
Calculates a transfer function between percentile of observed / simulated distributions. The transfer function is based on Hermite splines and a Gaussian kernel is fitted to observed data.
}
\author{
Sebastian Sippel
}
\references{
Sippel, S., Otto, F. E. L., Forkel, M., Allen, M. R., Guillod, B. P., Heimann, M., Reichstein, M., Seneviratne, S. I., Thonicke, K. & Mahecha, M. D. (2016) A novel bias correction methodology for climate impact simulations. Earth System Dynamics, 7, 71-88. doi:10.5194/esd-7-71-2016.
}

