% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.percentile.transfer.function.R
\name{resample.ensemble_no_replacement}
\alias{resample.ensemble_no_replacement}
\title{A function to derive a percentile transfer function}
\usage{
resample.ensemble_no_replacement(mod.grid.cell, transfer.function,
  sample.size = 1500, quantile.range = c(0, 1), search.radius)
}
\description{
A function to derive a percentile transfer function
}

