% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.percentile.transfer.function.R
\name{quality.control.ensemble}
\alias{quality.control.ensemble}
\title{A function to derive a percentile transfer function}
\usage{
quality.control.ensemble(transfer.function, percentile.ranges = seq(from = 0,
  to = 1, by = 0.1))
}
\description{
A function to derive a percentile transfer function
}

